restoring|at point of firebase addition
