+++ fedjobs/fedjobs/Controllers/UserProfileController.cs
@@ -30,7 +30,7 @@ namespace fedjobs.Controllers
         public IActionResult Post(UserProfile userProfile)
         {
             userProfile.CreateDateTime = DateTime.Now;
-            userProfile.UserTypeId = UserType.AUTHOR_ID;
+            
             _userProfileRepository.Add(userProfile);
             return CreatedAtAction(
                 nameof(GetUserProfile),
