+++ fedjobs/fedjobs/Repositories/BaseRepository.cs
@@ -1,5 +1,6 @@
-﻿using Microsoft.Data.SqlClient;
+﻿
 using Microsoft.Extensions.Configuration;
+using System.Data.SqlClient;
 
 namespace fedjobs.Repositories
 {
