+++ fedjobs/fedjobs/Repositories/IUserProfileRepository.cs
@@ -1,0 +1,10 @@
+﻿using fedjobs.Models;
+
+namespace fedjobs.Repositories
+{
+    public interface IUserProfileRepository
+    {
+        void Add(UserProfile userProfile);
+        UserProfile GetByFirebaseUserId(string firebaseUserId);
+    }
+}
