+++ fedjobs/fedjobs/Repositories/UserProfileRepository.cs
@@ -67,8 +67,7 @@ namespace fedjobs.Repositories
                     DbUtils.AddParameter(cmd, "@DisplayName", userProfile.DisplayName);
                     DbUtils.AddParameter(cmd, "@Email", userProfile.Email);
                     DbUtils.AddParameter(cmd, "@CreateDateTime", userProfile.CreateDateTime);
-                    DbUtils.AddParameter(cmd, "@ImageLocation", userProfile.ImageLocation);
-                    DbUtils.AddParameter(cmd, "@UserTypeId", userProfile.UserTypeId);
+                    
 
                     userProfile.Id = (int)cmd.ExecuteScalar();
                 }
