+++ fedjobs/fedjobs/Utils/DbUtils.cs
@@ -1,5 +1,6 @@
 ﻿using System;
-using Microsoft.Data.SqlClient;
+using System.Data.SqlClient;
+
 
 namespace fedjobs.Utils
 {
