+++ fedjobs/fedjobs/bin/Debug/net5.0/appsettings.json
@@ -1,0 +1,14 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
+    }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "DefaultConnection": "server=localhost\\SQLExpress;database=Tabloid;integrated security=true"
+  },
+  "FirebaseProjectId": "tabloid-c6ace"
+}
