+++ fedjobs/fedjobs/bin/Debug/net5.0/fedjobs.runtimeconfig.dev.json
@@ -1,0 +1,8 @@
+{
+  "runtimeOptions": {
+    "additionalProbingPaths": [
+      "C:\\Users\\Lanecw\\.dotnet\\store\\|arch|\\|tfm|",
+      "C:\\Users\\Lanecw\\.nuget\\packages"
+    ]
+  }
+}
