+++ fedjobs/fedjobs/client/src/components/Hello.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+export default function Hello() {
+  return (
+    <span style={{
+      position: "fixed",
+      left: 0,
+      right: 0,
+      top: "50%",
+      marginTop: "-0.5rem",
+      textAlign: "center",
+    }}>hello</span>
+  );
+}
